CA-233634: Use SPECS from xenserver-specs.git in xenbuilder build
This ensures that the xenbuilder iso and the transformer iso will
be building the same versions for all of the RPMs built in this
repo.

An additional step will now be required when pushing changes to
ely-staging: Both the xenserver-specs repo and this repo will need
a commit (for RPMs contained within this package).

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>