xenopsd: remove sub-packages for coverage builds for simulator and xenlight
The sub-packages for coverage builds are ill-defined and currently not
used. We need a global policy how to define analysis builds (coverage,
memory, time, ..). This commit keeps the coverage build of xenopsd that
is installed by the xenopsd-cov sub-package.

This brings xenopsd.spec in sync with
xenserver-specs/SPECS/xenopsd.spec.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>